Add images converter module
